Implemenent endpoints accessed by admin only
Add a book endpoint
[finishes #157067610]

Delete a book by it's id
[finishes #157067686]

Modify a book endpoint

[finishes #157067763]
