+++ app/admin/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+admin = Blueprint('admin', __name__)
+
+from . import views
